Create a protocol for updating mongo objects
